fruit-ninja-backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
scipy>=1.11.0
matplotlib>=3.8.0
pydantic>=2.4.0
psutil>=5.9.0
